refactor: provider 프로젝트 내 위치 변경
