+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:artificialsw_frontend/shell.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:artificialsw_frontend/features/puzzle/puzzle_completedlist.dart';
+import 'package:artificialsw_frontend/features/puzzle/puzzlelist_provider.dart';
 
 void main() {
   runApp(
